How to Evaluate an R Expression

The Job object defines several things:

The expression to run on a Worker process (expr).

What variables to make available during evaluation (vars).

How much time this task is allowed to take (timeout).

User-defined functions to run and when (hooks).

Result formatting (reformat).

Number of CPU cores to reserve (cpus).

Typically you won't need to call Job$new(). Instead, create a Queue and use <Queue>$run() to generate Job objects.

Active bindings

expr

Get the expression that will be run by this Job.

vars

Get or set the variables that will be placed into the expression's environment before evaluation.

scan

Get or set whether to scan for missing vars.

ignore

Get or set a character vector of variable names to NOT add to vars.

envir

Get or set the environment where missing vars can be found.

reformat

Get or set the function (job, output) for transforming raw callr. output to the Job's result.

catch

Get or set the signals to catch. output to the Job's result.

cpus

Get or set the number of CPUs to reserve for evaluating expr.

timeout

Get or set the time limits to apply to this Job.

proxy

Get or set the Job to proxy in place of running expr.

state

Get or set the Job's state (setting will trigger callbacks).

output

Get or set the Job's raw callr output (assigning to $output will change the Job's state to 'done').

result

Result of expr. Will block until Job is finished.

hooks

Currently registered callback hooks as a named list of functions.

is_done

Returns TRUE or FALSE depending on if the Job's result is ready.

uid

A short string, e.g. 'J16', that uniquely identifies this Job.

Methods

Public methods

Job$new()

Job$print()

Job$on()

Job$wait()

Job$stop()

Method new()

Creates a Job object defining how to run an expression on a background worker process.

Usage

Job$new(
  expr,
  vars = NULL,
  scan = FALSE,
  ignore = NULL,
  envir = parent.frame(),
  timeout = NULL,
  hooks = NULL,
  reformat = TRUE,
  catch = TRUE,
  cpus = 1L,
  ...
)

Arguments

expr

A call or R expression wrapped in curly braces to evaluate on a worker. Will have access to any variables defined by vars, as well as the Worker's globals, packages, and init configuration. See vignette('eval').

vars

A list of named variables to make available to expr during evaluation.

scan

Automatically add variables from envir to vars based on parsing/scanning expr. See vignette('scan').

ignore

A character vector of variable names that should NOT be added to vars by scan.

envir

Where to search for variables when scan = TRUE.

timeout

A named numeric vector indicating the maximum number of seconds allowed for each state the job passes through, or 'total' to apply a single timeout from 'submitted' to 'done'. Example: timeout = c(total = 2.5, running = 1). See vignette('stops').

hooks

A list of functions to run when the Job state changes, of the form hooks = list(created = function (job) {...}, done = ~{...}). See vignette('hooks').

reformat

Set reformat = FALSE to return the entire callr output (<Job>$output), or reformat = function (job) to use a function of your own to post-process the output from callr. See vignette('results').

catch

What types of conditions to catch, e.g. c('interrupt', 'error', 'warning'). Or TRUE / FALSE to catch all/none. A caught condition will be returned by <Job>$result, otherwise the condition will be signaled with stop(<condition>) when <Job>$result is called. See vignette('results').

cpus

How many CPU cores to reserve for this Job. The Queue uses this number to limit the number of simultaneously running Jobs; it does not prevent a Job from using more CPUs than reserved.

...

Arbitrary named values to add to the returned Job object.

Returns

A Job object.

Method print()

Print method for a Job.

Usage

Job$print(...)

Arguments

...

Arguments are not used currently.

Returns

This Job, invisibly.

Method on()

Attach a callback function.

Usage

Job$on(state, func)

Arguments

state

The Job state that will trigger this function. Typically one of:

'*' - Every time the state changes.

'.next' - Only one time, the next time the state changes.

'created' - After Job$new() initialization.

'submitted' - After <Job>$queue is assigned.

'queued' - After stop_id and copy_id are resolved.

'dispatched' - After <Job>$worker is assigned.

'starting' - Before evaluation begins.

'running' - After evaluation begins.

'done' - After <Job>$output is assigned.

Custom states can also be set/used.

func

A function that accepts a Job object as input. You can call <Job>$stop() or edit its values and the changes will be persisted (since Jobs are reference class objects). You can also edit/stop other queued jobs by modifying the Jobs in <Job>$queue$jobs. Return value is ignored.

Returns

A function that when called removes this callback from the Job.

Method wait()

Blocks until the Job enters the given state.

Usage

Job$wait(state = "done")

Arguments

state

The Job state that will trigger this function. Typically one of:

'*' - Every time the state changes.

'.next' - Only one time, the next time the state changes.

'created' - After Job$new() initialization.

'submitted' - After <Job>$queue is assigned.

'queued' - After stop_id and copy_id are resolved.

'dispatched' - After <Job>$worker is assigned.

'starting' - Before evaluation begins.

'running' - After evaluation begins.

'done' - After <Job>$output is assigned.

Custom states can also be set/used.

Returns

This Job, invisibly.

Method stop()

Stop this Job. If the Job is running, the worker process will be rebooted.

Usage

Job$stop(reason = "job stopped by user")

Arguments

reason

A message to include in the 'interrupt' condition object that will be returned as the Job's result.

Returns

This Job, invisibly.